refactor: update motor settings and improve rotation handling
- Changed MOTOR_STEPS type from int64_t to uint16_t for better memory usage.
- Introduced MOTOR_ROTATIONS_LIMIT to prevent overflow in rotation calculations.
- Updated motor component logic to use constrained rotations and adjusted step calculations.
- Modified data structures to use uint32_t for rotations to accommodate larger values.
- Enhanced validation and input handling in the UI for motor settings, including steps per pulse.
- Updated protocol definitions to reflect changes in rotation data types for consistency.

+++ arduino/lib/components/motor.hpp
@@ -37,7 +37,7 @@ namespace components
             {
                 uint8_t slave;
                 bool direction;
-                uint16_t rotations;
+                uint32_t rotations;
             } MOVE_DATA;
 
             typedef struct __attribute__((packed))
